+++ test/src/analyzers/lint_analyzer/rules/rules_list/no_blank_line_before_single_return/examples/fix.after.dart
@@ -42,5 +42,9 @@ int f(int a) {
     return a + 8;
   }
 
+  if (a > 9) { // trailing brace comment
+    return a + 9;
+  }
+
   return a;
 }