+++ test/src/analyzers/lint_analyzer/rules/rules_list/no_blank_line_before_single_return/examples/fix.before.dart
@@ -54,5 +54,10 @@ int f(int a) {
     return a + 8;
   }
 
+  if (a > 9) { // trailing brace comment
+
+    return a + 9;
+  }
+
   return a;
 }